+++ CHANGELOG.md
@@ -86,6 +86,10 @@ Tracing tool:
   added to the active stack viewer by selecting the respective remote project
   from the "More tools" menu item "Remote data".
 
+- To open a remote CATMAID instance at the same location as the current view,
+  find the remote project in the "Remote data" section of the "More tools" menu,
+  and click on "Open remote view".
+
 - The marker type used fore remote tracing data can be configured through the
   Settings Widget in the Tracing Overlay section, option "Remote node marker". A
   ring is used by default.